+++ examples/custom_checkers/c9103_invalid_name_checker.py
@@ -1,0 +1,11 @@
+"""Examples for C9103 pep8-name-violation."""
+
+not_upper_for_const = "CONSTANT"  # pep8-name-violation
+
+
+def NotSnakeCase():  # pep8-name-violation
+    pass
+
+
+class not_pascal_case:  # pep8-name-violation
+    pass